requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
ipykernel
beautifulsoup4
lxml
selenium
tbselenium
SQLAlchemy
mysql-connector-python
SQLAlchemy-Utils
pandas
numpy
pyarrow
scikit-learn==1.4.1.post1
matplotlib
typeguard
seaborn
xgboost
stem
hydra-core
pyyaml
pactools